Intelligent Resume Analyzer and Career Advisor/requirements.txt
langchain
langchain_openai
streamlit
python-dotenv
pypdf